requirements.txt
langchain 
wikipedia 
langchain-community
python-dotenv
pydantic
duckduckgo-search
langchain-google-genai
gradio
requests
beautifulsoup4